+++ ntex-tls/src/openssl/accept.rs
@@ -1,12 +1,12 @@
 use std::task::{Context, Poll};
-use std::{error::Error, future::Future, marker::PhantomData, pin::Pin};
+use std::{error::Error, marker::PhantomData};
 
 use ntex_io::{Filter, FilterFactory, Io, Layer};
 use ntex_service::{Service, ServiceCtx, ServiceFactory};
-use ntex_util::{future::Ready, time::Millis};
+use ntex_util::time::Millis;
 use tls_openssl::ssl::SslAcceptor;
 
-use crate::counter::{Counter, CounterGuard};
+use crate::counter::Counter;
 use crate::MAX_SSL_ACCEPT_COUNTER;
 
 use super::{SslAcceptor as IoSslAcceptor, SslFilter};
@@ -58,12 +58,10 @@ impl<F: Filter, C: 'static> ServiceFactory<Io<F>, C> for Acceptor<F> {
     type Error = Box<dyn Error>;
     type Service = AcceptorService<F>;
     type InitError = ();
-    type Future<'f> = Ready<Self::Service, Self::InitError>;
 
-    #[inline]
-    fn create(&self, _: C) -> Self::Future<'_> {
+    async fn create(&self, _: C) -> Result<Self::Service, Self::InitError> {
         MAX_SSL_ACCEPT_COUNTER.with(|conns| {
-            Ready::Ok(AcceptorService {
+            Ok(AcceptorService {
                 acceptor: self.acceptor.clone(),
                 conns: conns.clone(),
                 _t: PhantomData,
@@ -85,9 +83,7 @@ pub struct AcceptorService<F> {
 impl<F: Filter> Service<Io<F>> for AcceptorService<F> {
     type Response = Io<Layer<SslFilter, F>>;
     type Error = Box<dyn Error>;
-    type Future<'f> = AcceptorServiceResponse<F>;
 
-    #[inline]
     fn poll_ready(&self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
         if self.conns.available(cx) {
             Poll::Ready(Ok(()))
@@ -96,30 +92,12 @@ impl<F: Filter> Service<Io<F>> for AcceptorService<F> {
         }
     }
 
-    #[inline]
-    fn call<'a>(&'a self, req: Io<F>, _: ServiceCtx<'a, Self>) -> Self::Future<'a> {
-        AcceptorServiceResponse {
-            _guard: self.conns.get(),
-            fut: self.acceptor.clone().create(req),
-        }
-    }
-}
-
-pin_project_lite::pin_project! {
-    pub struct AcceptorServiceResponse<F>
-    where
-        F: Filter,
-    {
-        #[pin]
-        fut: <IoSslAcceptor as FilterFactory<F>>::Future,
-        _guard: CounterGuard,
-    }
-}
-
-impl<F: Filter> Future for AcceptorServiceResponse<F> {
-    type Output = Result<Io<Layer<SslFilter, F>>, Box<dyn Error>>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        self.project().fut.poll(cx)
+    async fn call(
+        &self,
+        req: Io<F>,
+        _: ServiceCtx<'_, Self>,
+    ) -> Result<Self::Response, Self::Error> {
+        let _guard = self.conns.get();
+        self.acceptor.clone().create(req).await
     }
 }