+++ ntex-connect/CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [1.0.0] - 2024-01-0x
+
+* Use "async fn" in trait for Service definition
+
 ## [0.3.4] - 2023-12-14
 
 * Better io tag handling